chore: moving all command files into a new `commands` directory
