+++ src/lib/commands/add.js
