+++ src/lib/commands/docs.js
