+++ src/lib/commands/generate.js
