+++ src/lib/commands/help.js
