+++ src/lib/commands/login.js
