+++ src/lib/commands/validate.js
